Was not respecting keepLog

+++ src/main/java/jenkins/plugins/purgejobhistory/PurgeJobHistory.java
@@ -46,6 +46,12 @@ public class PurgeJobHistory extends CLICommand {
     public boolean reset = false;
 
     /**
+     * Follows the progress of the operation
+     */
+    @Option(name = "-f", usage = "Force deleting all builds.")
+    public boolean force = false;
+
+    /**
      * The source item.
      */
     @Argument(metaVar = "JOB", usage = "Name of the job whose history should be purged", required = true)
@@ -76,6 +82,17 @@ public class PurgeJobHistory extends CLICommand {
      * @throws IOException if something went wrong.
      */
     public static void purge(Job<?, ?> job, boolean resetNextBuildNumber) throws IOException {
+        purge(job, resetNextBuildNumber, false);
+    }
+    /**
+     * Purges the build history of the specified job.
+     *
+     * @param job                  the job to purge
+     * @param resetNextBuildNumber {@code true} if the next build number should be reset to {@code 1} after the purge
+     * @throws IOException if something went wrong.
+     * @since 1.1
+     */
+    public static void purge(Job<?, ?> job, boolean resetNextBuildNumber, boolean force) throws IOException {
         ACL lastACL = null;
         for (Run<?, ?> run : new ArrayList<Run<?, ?>>(job.getBuilds())) {
             ACL acl = run.getACL();
@@ -85,9 +102,12 @@ public class PurgeJobHistory extends CLICommand {
                 acl.checkPermission(Run.DELETE);
                 lastACL = acl;
             }
+            if (!force && run.isKeepLog()) {
+                continue;
+            }
             run.delete();
         }
-        if (resetNextBuildNumber) {
+        if (resetNextBuildNumber && job.getLastBuild() == null) {
             job.updateNextBuildNumber(1);
         }
     }
